pyproject.toml
[project]
name = "crrl"
version = "0.1.0"
description = "Finetuning LLMs with RL to improve program repair capabilities"
readme = "README.md"
requires-python = ">=3.11,<3.13"
dependencies = [
    "datasets>=4.4.1",
    "gitpython>=3.1.45",
    "hydra-core>=1.3.2",
    "mini-swe-agent>=1.17.0",
    "nano-agent",
    "wandb>=0.23.0",
]

[project.optional-dependencies]
# vLLM is required, but install is brittle needing torch to be installed first
vllm = ["vllm==0.11.0; platform_system != 'Darwin'"]
# Flash attention packages - flash-attn needed by transformers for training, flashinfer for vLLM
flash = [
    "flash-attn==2.8.3; platform_system != 'Darwin'",
    "flashinfer-python==0.5.0; platform_system != 'Darwin'",
]
dev = ["pytest>=8.3.4", "matplotlib>=3.10.0", "scikit-learn>=1.6.1"]
# Different agents
# aider = ["aider-chat @ git+https://github.com/BjarniHaukur/aider.git"]
# sweagent = ["sweagent @ git+https://github.com/BjarniHaukur/SWE-agent.git"]

[tool.uv]
no-build-isolation-package = ["flash-attn", "flashinfer-python"]  
[tool.uv.sources]
nano-agent = { git = "https://github.com/ASSERT-KTH/nano-agent", branch = "andre/feat/remote-env" }
